Adding register

+++ proyecto3_discord/base/views.py
@@ -1,7 +1,10 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from .models import Room, Topic
 from .forms import RoomForm
@@ -13,8 +16,14 @@ from .forms import RoomForm
 ] """
 
 def login_page(request):
+    
+    page= 'login'
+    
+    if request.user.is_authenticated:
+        return redirect('home')
+    
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -30,13 +39,28 @@ def login_page(request):
             messages.error(request, 'Wrong credentials, please try one more time')  
  
                
-    context = {}
+    context = {'page': page}
     return render(request, 'base/login_register.html', context)
 
 
 def logout_user(request):
     logout(request)
     return redirect('home')
+
+def register_user(request):
+    form = UserCreationForm()
+    
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error has occurred during registration')
+    return render(request, 'base/login_register.html', {'form': form})
 
 def home(request):
     query_param = request.GET.get('q') if request.GET.get('q') else ''
@@ -55,6 +79,7 @@ def room(request, pk):
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+@login_required(login_url='login')
 def create_room(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -66,9 +91,13 @@ def create_room(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html',context)
 
+@login_required(login_url='login')
 def update_room(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to update this room!')
     
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -79,8 +108,13 @@ def update_room(request, pk):
     context = {'form':form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url='login')
 def delete_room(request, pk):
     room = Room.objects.get(id=pk)
+    
+    if request.user != room.host:
+        return HttpResponse('You are not allowed to delete this room!')
+    
     if request.method == 'POST':
         room.delete()
         return redirect('home')
